Organized codebase.
McRaw is going to be a rather big project, and it would be prudent to
separate out the files.  I haven't figured out how to do import
statements successfully (at least not working with how I want the
project set up), so all I do is execfile on each of them.